ScottPlot5: refactor Axes
