+++ dev/ScottPlot5/ScottPlot/Axes/IAxis.cs
@@ -5,6 +5,7 @@ namespace ScottPlot.Axes;
 public interface IAxis
 {
     Edge Edge { get; }
+    Orientation Orientation { get; }
     TextLabel Label { get; }
     PixelSize GetSize(ICanvas canvas);
     void Draw(ICanvas canvas, PlotInfo info);
